improved code plugin section

+++ frontend/src/components/plugins/utils/utils.jsx
@@ -25,8 +25,12 @@ import {
   useOrganizationStore,
   usePluginConfigurationStore,
 } from "../../../stores";
-import {ANALYZERS_CONFIG_URI, CONNECTORS_CONFIG_URI, VISUALIZERS_CONFIG_URI} from "../../../constants/api";
-import {ANALYZER, CONNECTOR, VISUALIZER} from "../../../constants/constants";
+import {
+  ANALYZERS_CONFIG_URI,
+  CONNECTORS_CONFIG_URI,
+  VISUALIZERS_CONFIG_URI,
+} from "../../../constants/api";
+import { pluginType } from "../../../constants/constants";
 
 const { checkPluginHealth } = usePluginConfigurationStore.getState();
 
@@ -219,13 +223,13 @@ function PluginHealthSpinner() {
   return <Spinner type="ripple" size="sm" className="text-darker" />;
 }
 
-export function PluginHealthCheckButton({ pluginName, pluginType }) {
+export function PluginHealthCheckButton({ pluginName, pluginType_ }) {
   const [isLoading, setIsLoading] = React.useState(false);
   const [isHealthy, setIsHealthy] = React.useState(undefined);
 
   const onClick = async () => {
     setIsLoading(true);
-    const status = await checkPluginHealth(pluginType, pluginName);
+    const status = await checkPluginHealth(pluginType_, pluginName);
     setIsHealthy(status);
     setIsLoading(false);
   };
@@ -275,18 +279,17 @@ export function OrganizationPluginStateToggle({
   if (!organization.name) title = "You're not a part of any organization";
   else if (!isUserOwner)
     title = `You're not an owner of your organization - ${organization.name}`;
-  let baseUrl = ""
-  if (type === ANALYZER){
-      baseUrl = ANALYZERS_CONFIG_URI
-  } else if (type === CONNECTOR ){
-      baseUrl = CONNECTORS_CONFIG_URI
-  }else if (type === VISUALIZER){
-      baseUrl = VISUALIZERS_CONFIG_URI
+  let baseUrl = "";
+  if (type === pluginType.ANALYZER) {
+    baseUrl = ANALYZERS_CONFIG_URI;
+  } else if (type === pluginType.CONNECTOR) {
+    baseUrl = CONNECTORS_CONFIG_URI;
+  } else if (type === pluginType.VISUALIZER) {
+    baseUrl = VISUALIZERS_CONFIG_URI;
   }
 
   const onClick = async () => {
-    if (disabled)
-      await axios.delete(`${baseUrl}/${pluginName}/organization`);
+    if (disabled) await axios.delete(`${baseUrl}/${pluginName}/organization`);
     else await axios.post(`${baseUrl}/${pluginName}/organization`);
     fetchAllOrganizations();
     refetch();
@@ -328,5 +331,5 @@ PluginVerificationIcon.propTypes = {
 
 PluginHealthCheckButton.propTypes = {
   pluginName: PropTypes.string.isRequired,
-  pluginType: PropTypes.oneOf(["analyzer", "connector"]).isRequired,
+  pluginType_: PropTypes.oneOf(["analyzer", "connector"]).isRequired,
 };
